Updates folder structure, fixes #350
