+++ generators/all/index.js
