+++ generators/app/index.js
