+++ generators/app/templates/Gruntfile.js
