+++ generators/app/templates/_bower.json
