+++ generators/app/templates/_package.json
