+++ generators/app/templates/app.coffee
