+++ generators/app/templates/app.js
