+++ generators/app/templates/app/favicon.ico
