+++ generators/app/templates/bowerrc
