+++ generators/app/templates/editorconfig
