+++ generators/app/templates/gitattributes
