+++ generators/app/templates/gitignore
