+++ generators/app/templates/jshintrc
