+++ generators/app/templates/main.scss
