+++ generators/app/templates/requirejs_app.coffee
