+++ generators/app/templates/requirejs_app.js
