+++ generators/templates/coffeescript/collection.coffee
