+++ generators/templates/coffeescript/model.coffee
