+++ generators/templates/coffeescript/requirejs/collection.coffee
