+++ generators/templates/coffeescript/requirejs/model.coffee
