+++ generators/templates/coffeescript/requirejs/router.coffee
