+++ generators/templates/coffeescript/requirejs/view.coffee
