+++ generators/templates/coffeescript/router.coffee
