+++ generators/templates/coffeescript/view.coffee
