+++ generators/templates/coffeescript/view.ejs
