+++ generators/templates/collection.js
