+++ generators/templates/model.js
