+++ generators/templates/requirejs/collection.js
