+++ generators/templates/requirejs/model.js
