+++ generators/templates/requirejs/router.js
