+++ generators/templates/requirejs/view.js
