+++ generators/templates/router.js
