+++ generators/templates/view.ejs
