+++ generators/templates/view.js
